+++ CsharpExtras/Map/Sparse/TwoDimensional/ISparseArray2D.cs
@@ -73,5 +73,23 @@ namespace CsharpExtras.Map.Sparse.TwoDimensional
         /// <param name="leftmostRow">The row coordinate of the top-left corner of the rectangle to which to write</param>
         /// <param name="topMostColumn">The column coordinate of the top-left corner of the rectangle to which to write</param>
         void SetArea(TVal[,] area, int leftmostRow, int topMostColumn);
+
+        /// <summary>
+        /// Zips this array with another one to produce a new array.
+        /// </summary>
+        /// <typeparam name="TOther">The type of elements in the other array</typeparam>
+        /// <typeparam name="TResult">The type of elements in the resultant array</typeparam>
+        /// <param name="zipper">A function which, given an element from this array and the other array,
+        /// returns an element of the resultant array.</param>
+        /// <param name="other">The other array with which to zip this one</param>
+        /// <param name="defaultVal">The default value for the new sparse array</param>
+        /// <param name="validationFunction">Validates indices in the resultant array. If any zipped indices are invalid,
+        /// the zip operation will throw an exception.</param>
+        /// <returns>A new sparse 2D array, the used indices of which will be the union of used indices for this array and the other array
+        /// and the values of which will be the result of applying the zipper function at the corresponding indices in this array and the other array</returns>
+        /// <returns></returns>
+        ISparseArray2D<TResult> Zip<TOther, TResult>(Func<TVal, TOther, TResult> zipper,
+            ISparseArray2D<TOther> other, TResult defaultVal,
+            Func<NonnegativeInteger, int, bool> validationFunction);
     }
 }