+++ CHANGELOG.md
@@ -43,7 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added general-purpose Sparse Array type for any amount of dimensions with builder and some basic functions (#108)
 - Added sparse 2D array type for 2D-specific operations (#120)
 - Added a shift operator for sparse arrays (#122)
-- Added a pre-access wrapper, which wraps an object and ensures an action is run before any access to that object (#131_
+- Added a pre-access wrapper, which wraps an object and ensures an action is run before any access to that object (#131)
+- Added zip methods for sparse array and sparse array 2D (#132)
 
 ### Added
 - Support for writing a 1D one-based array to a 2D one-based array (#9)